Core/GameObjects: Changed gameobject artkit type to uint
Port From (https://github.com/TrinityCore/TrinityCore/commit/bb3b2cf4cb4897a7dd087e2409a250c887c3fb9a)

+++ Source/Game/Chat/Commands/GameObjectCommands.cs
@@ -753,7 +753,7 @@ namespace Game.Chat
                         obj.SetGoType((GameObjectTypes)objectState);
                         break;
                     case 2:
-                        obj.SetGoArtKit((byte)objectState);
+                        obj.SetGoArtKit(objectState);
                         break;
                     case 3:
                         obj.SetGoAnimProgress(objectState);
